add page idea

+++ app/Models/IdeaMedia.php
@@ -11,42 +11,43 @@ class IdeaMedia extends BaseModel
 
     protected $table = 'idea_media';
     public $attachmentable = [
-        'url' =>[
+        'url' => [
                 'path' => [
                     'crop' => 'storage/attachments/crops',
                     'storage' => 'attachments'
-                ],
+                    ],
+                'filename' => 'filename',
                 'default' => 'assets/images/idea.jpg'
-        ]
+            ]
     ];
     protected $fillable = [
-        'idea_id', 'url', 'type'
+        'idea_id', 'url', 'type', 'filename'
     ];
 
     public static function boot()
     {
         parent::boot();
         
-        static::bootAttachableTrait();
         static::creating(function($ideaMedia){
-            if (!is_string($ideaMedia->url)) {
-                $mime = mime_content_type($ideaMedia->url);
-                if(strstr($mime, "video/")){
-                    $ideaMedia->type = 'video';
-                }else if(strstr($mime, "image/")){
-                    $ideaMedia->type = 'image';
-                } else {
-                    $ideaMedia->type = 'file';
-                }
+            $mime = mime_content_type($ideaMedia->url);
+            if(strstr($mime, "video/")){
+                $ideaMedia->type = 'video';
+            }else if(strstr($mime, "image/")){
+                $ideaMedia->type = 'image';
+            } else {
+                $ideaMedia->type = 'file';
             }
         });
         static::deleted(function($ideaMedia){
-            $name = basename($ideaMedia->url);
-            array_map('unlink', glob($ideaMedia->attachmentable['url']['path']['crop']."/".$name.'-*'));
-            if ($ideaMedia->url != $ideaMedia->attachmentable['url']['default']) {
-                unlink($ideaMedia->url);
+            if(is_null(static::withTrashed()->find($ideaMedia->id))) {
+                $name = basename($ideaMedia->url);
+                array_map('unlink', glob($ideaMedia->attachmentable['url']['path']['crop']."/".$name.'-*'));
+                if ($ideaMedia->url != $ideaMedia->attachmentable['url']['default']) {
+                    unlink($ideaMedia->url);
+                }
             }
         });
+        static::bootAttachableTrait();
     }
 
     public function idea()
